created NavListItem component
