+++ src/components/navListItem.tsx
@@ -1,0 +1,33 @@
+import { useNavigate } from "react-router-dom";
+import { ListItem, ListItemButton, ListItemIcon, ListItemText } from "@mui/material";
+
+export default function NavListItem(props: any) {
+  const { open, name, linkTo, children } = props;
+  const navigate = useNavigate();
+
+  return (
+    <ListItem key={name} disablePadding sx={{ display: "block" }}>
+      <ListItemButton
+        sx={{
+          minHeight: 48,
+          justifyContent: open ? "initial" : "center",
+          px: 2.5,
+        }}
+        onClick={() => {
+          navigate(`/app/${linkTo}`);
+        }}
+      >
+        <ListItemIcon
+          sx={{
+            minWidth: 0,
+            mr: open ? 3 : "auto",
+            justifyContent: "center",
+          }}
+        >
+          {children}
+        </ListItemIcon>
+        <ListItemText primary={name} sx={{ opacity: open ? 1 : 0 }} />
+      </ListItemButton>
+    </ListItem>
+  );
+}
